refactor(ContributionFlow): Update files names
