+++ src/server/pages.js
@@ -58,13 +58,17 @@ pages
 // ---------------
 
 // Special route to force Legacy Flow
-pages.add('orderCollectiveLegacyForce', '/:collectiveSlug/:verb(donate|pay|contribute|order)/legacy', 'createOrder');
+pages.add(
+  'orderCollectiveLegacyForce',
+  '/:collectiveSlug/:verb(donate|pay|contribute|order)/legacy',
+  'createOrderLegacy',
+);
 
 // Legacy create order route. Deprectated on 2019-02-12
 pages.add(
   'orderCollectiveTier',
   '/:collectiveSlug/:verb(order)/:tierId/:amount(\\d+)?/:interval(month|monthly|year|yearly)?',
-  'createOrderNewFlow',
+  'createOrder',
 );
 
 // New Routes -> New flow
@@ -72,12 +76,12 @@ pages
   .add(
     'orderCollectiveNew',
     '/:collectiveSlug/:verb(donate|pay|contribute|order)/:step(contributeAs|details|payment)?',
-    'createOrderNewFlow',
+    'createOrder',
   )
   .add(
     'orderCollectiveTierNew',
     '/:collectiveSlug/:verb(donate|pay|contribute|order)/tier/:tierId-:tierSlug?/:step(contributeAs|details|payment)?',
-    'createOrderNewFlow',
+    'createOrder',
   )
   .add('orderCollectiveNewSuccess', '/:collectiveSlug/:verb(donate|pay|contribute|order)/success', 'orderSuccess')
   .add(
@@ -90,11 +94,11 @@ pages
 pages.add(
   'orderCollective',
   '/:collectiveSlug/:verb(donate|pay|contribute|order)/:amount(\\d+)?/:interval(month|monthly|year|yearly)?/:description?',
-  'createOrderNewFlow',
+  'createOrder',
 );
 
 // New contribution flow not applied to events yet
-pages.add('orderEventTier', '/:collectiveSlug/events/:eventSlug/order/:TierId', 'createOrder');
+pages.add('orderEventTier', '/:collectiveSlug/events/:eventSlug/order/:TierId', 'createOrderLegacy');
 
 // Pledges
 // -------
